move posts to own page, add placeholder homepage
